api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
    #   starlette
    #   watchfiles
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.5
    # via requests
click==8.3.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
cryptography==46.0.5
    # via google-auth
docstring-parser==0.17.0
    # via google-cloud-aiplatform
fastapi==0.135.1
    # via -r requirements.in
google-api-core[grpc]==2.30.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth==2.49.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-resource-manager
    #   google-cloud-storage
    #   google-genai
google-cloud-aiplatform[all]==1.71.1
    # via
    #   -r requirements.in
    #   vertexai
google-cloud-bigquery==3.40.1
    # via
    #   -r requirements.in
    #   google-cloud-aiplatform
google-cloud-core==2.5.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.24.0
    # via -r requirements.in
google-cloud-resource-manager==1.16.0
    # via google-cloud-aiplatform
google-cloud-storage==2.19.0
    # via
    #   -r requirements.in
    #   google-cloud-aiplatform
google-crc32c==1.8.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.8.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.73.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.3
    # via google-cloud-resource-manager
grpcio==1.78.0
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httptools==0.7.1
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
numpy==2.4.2
    # via shapely
packaging==26.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
proto-plus==1.27.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-firestore
    #   google-cloud-resource-manager
protobuf==5.29.6
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-firestore
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
    #   google-cloud-aiplatform
    #   google-genai
pydantic-core==2.41.5
    # via pydantic
python-dateutil==2.9.0.post0
    # via google-cloud-bigquery
python-dotenv==1.2.2
    # via uvicorn
python-multipart==0.0.22
    # via -r requirements.in
pyyaml==6.0.3
    # via uvicorn
requests==2.32.5
    # via
    #   google-api-core
    #   google-auth
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   google-genai
rsa==4.9.1
    # via google-auth
shapely==2.1.2
    # via google-cloud-aiplatform
six==1.17.0
    # via python-dateutil
starlette==0.52.1
    # via fastapi
typing-extensions==4.15.0
    # via
    #   fastapi
    #   google-genai
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
urllib3==2.6.3
    # via requests
uvicorn[standard]==0.41.0
    # via -r requirements.in
vertexai==1.71.1
    # via -r requirements.in
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via
    #   google-genai
    #   uvicorn
python-dotenv==1.2.2
    # via uvicorn
